// src/types/enum/Occurrence.ts
export enum Occurrence {
  DUPLICATION = "Duplication",
  RESPONSIBILITY = "Responsibility",
  MEASURED_SMELL = "Measured Smells",
  DATA = "Data",
  UNNECESSARY_COMPLEXITY = "Unnecessary Complexity",
  INTERFACES = "Interfaces",
  NAMES = "Names",
  MESSAGE_CALLS = "Message Calls",
  CONDITIONAL_LOGIC = "Conditional Logic",
  UNKNOWN = "Unknown",
}